+++ backend/server.js
@@ -1,18 +1,19 @@
 const express = require("express");
-const bodyParser = require("body-parser");
-const getBusStops = require("./middleware/getBusStops");
+const getDepartures = require("./middleware/getDepartures");
 
 const app = express();
+const port = 3000;
 
 app.use(express.json());
-
-const port = 3000;
 
 app.get("/", (request, response) => {
   response.send("Hello World Happy Days");
 });
 
-app.get("/departureBoards", getBusStops.getBusStopsMiddleware);
+app.get(
+  "/departureBoards",
+  getDepartures.getNextFiveBusDepartureTimesForGivenPostcode
+);
 
 app.listen(port, () => {
   console.log(`Listening on ${port}`);
