+++ exercises/partOneBusTimes.js
@@ -1,5 +1,5 @@
 const axios = require("axios");
-const apiCredentials = require("./apiCredentials");
+const apiCredentials = require("../apiCredentials");
 
 const BASE_URL = "https://api.tfl.gov.uk";
 const service = "/StopPoint";
@@ -11,7 +11,6 @@ const constructPath = (stopCode) => {
 
 const getLiveTimes = async (stopCode) => {
   const path = constructPath(stopCode);
-  // console.log({path});
 
   try {
     return await axios.get(path);
@@ -25,9 +24,9 @@ const getBuses = async (stopCode) => {
   try {
     const liveTimes = await getLiveTimes(stopCode);
     const fiveBuses = liveTimes.data.slice(0, 5);
-    // fiveBuses.forEach((el) => {
-    //   console.log(el.lineId, el.destinationName, el.timeToStation);
-    // });
+    fiveBuses.forEach((el) => {
+      console.log(el.lineId, el.destinationName, el.timeToStation);
+    });
   } catch (error) {
     console.log("second block", error);
   }
