--- jobsToDo.js
@@ -1,11 +1,0 @@
-// TODO
-// Create a function which takes the type of url we want and the parameters and returns the correct path
-
-// const constructPath = (stopCode) => {
-//     return `${BASE_URL}${service}${stopCode}/Arrivals?app_id=${apiCredentials.id}&app_key=${apiCredentials.key}`;
-//   };
-
-// const constructPath = (lat, lon) => {
-//     return `${BASE_URL}${service}?lat=${lat}&lon=${lon}&stoptypes=${stopTypes}&radius=${radius}&app_id=${apiCredentials.id}&app_key=${apiCredentials.key}`;
-//   };
-
